REFACTOR: Separated language modal into separate component
